+++ src/lib.rs
@@ -415,6 +415,7 @@ pub mod element_decode;
 pub mod extension_payload;
 pub mod filterbank;
 pub mod gain_control_data;
+pub mod hcr;
 pub mod ics_body;
 pub mod ics_info;
 pub mod intensity_stereo;